Fix three type conversion warnings
Fixed sign conversion warnings in PBHistorySearchController and PBGitSidebarController:
- Added NSUInteger casts for selectedRow parameter in indexGreaterThanIndex/indexLessThanIndex calls
- Added NSInteger casts for count method returns in outlineView numberOfChildrenOfItem method

+++ Classes/Controllers/PBHistorySearchController.m
@@ -169,9 +169,9 @@
 
 	NSUInteger currentResult = NSNotFound;
 	if (direction == kGitXSearchDirectionNext)
-		currentResult = [results indexGreaterThanIndex:selectedRow];
+		currentResult = [results indexGreaterThanIndex:(NSUInteger)selectedRow];
 	else
-		currentResult = [results indexLessThanIndex:selectedRow];
+		currentResult = [results indexLessThanIndex:(NSUInteger)selectedRow];
 
 	if (currentResult == NSNotFound) {
 		if (direction == kGitXSearchDirectionNext)
